# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
paths:
  /tasks:
    get:
      summary: Get all tasks
      tags:
        - tasks
      responses:
        "200":
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskResponse"
    post:
      summary: Create a new task
      tags:
        - tasks
      requestBody:
        description: The task to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreatePatchBody"
      responses:
        "201":
          description: The created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
  /tasks/{id}:
    get:
      summary: Get a task by ID
      tags:
        - tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
    patch:
      summary: Patch a task by ID
      tags:
        - tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: The task to patch
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreatePatchBody"
      responses:
        "200":
          description: The task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
    delete:
      summary: Delete a task by ID
      tags:
        - tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The task was deleted

components:
  schemas:
    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          format: uint
        title:
          type: string
        description:
          type: string
        completed:
          type: boolean
    TaskCreatePatchBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        completed:
          type: boolean